+++ src/backend/adminServer.ts
@@ -31,8 +31,9 @@ AdminServerModuleLoader.loadAdminModules(app, serverConfig);
 const bindIp = serverConfig.adminBackendConfig.bindIp ? serverConfig.adminBackendConfig.bindIp : '127.0.0.1';
 const tcpBacklog = serverConfig.adminBackendConfig.tcpBacklog ? serverConfig.adminBackendConfig.tcpBacklog : 511;
 app.listen(serverConfig.adminBackendConfig.port, bindIp,  tcpBacklog, function () {
-    console.log('MyTB app listening on ip/port/tcpBacklog', bindIp, serverConfig.adminBackendConfig.port, tcpBacklog);
+    console.log('MyShp app listening on ip/port/tcpBacklog', bindIp, serverConfig.adminBackendConfig.port, tcpBacklog);
 
+    console.log('Debug-level level/log/trace', debug, !(!debug), !(!debug || debug === true || parseInt(debug, 10) < 1));
     if (!debug) {
         console.log = function() {};
     }
@@ -40,5 +41,8 @@ app.listen(serverConfig.adminBackendConfig.port, bindIp,  tcpBacklog, function (
         console.trace = function() {};
         console.debug = function() {};
     }
+
+    console.log('log enabled');
+    console.trace('trace enabled');
 });