+++ src/frontend/environments/environment.prod.ts
@@ -1,4 +1,5 @@
 import {AppEnvironment} from './app-environment';
+import {DataMode} from '../shared/commons/model/datamode.enum';
 
 export const environment: AppEnvironment = {
     production: true,
@@ -11,7 +12,11 @@ export const environment: AppEnvironment = {
     sdocMaxItemsPerAlbum: 2000,
     backendApiBaseUrl: 'http://localhost:4102/api/v1/',
     cookieLawSeenName: 'cookieLawSeenV20180525',
-    trackingProviders: [] // Angulartics2Piwik
+    trackingProviders: [], // Angulartics2Piwik
+    hideInternalDescLinks: false,
+    hideInternalImages: false,
+    startDataMode: DataMode.BACKEND,
+    availableDataModes: [DataMode.BACKEND]
 };
 
 // unset logger